% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manta_vcf2bedpe.R
\name{bnd_matching}
\alias{bnd_matching}
\title{match breakends}
\usage{
bnd_matching(id, bnd)
}
\arguments{
\item{id}{mate id to pair breakpoints}

\item{bnd}{data table of breakend SVs}
}
\value{
SV data table with matched breakpoints
}
\description{
Pairs mate ids from vcf and constructs bedpe row for the SV
}
\details{
Read mate pairing for bedpe construction, internally used
}
\keyword{internal}
