% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_sine_annotation.R
\name{check_reformat}
\alias{check_reformat}
\title{reformat for matching}
\usage{
check_reformat(i, df)
}
\arguments{
\item{i}{iteration value}

\item{df}{bedpe to be passed}
}
\value{
data.table that is ordered with lower position first
}
\description{
Reorders foor acccurate distance calculation
}
\details{
Check format of bedpe
}
\keyword{internal}
