% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_sine_annotation.R
\name{closest_line_sine}
\alias{closest_line_sine}
\title{Annotate LINE and SINE elements}
\usage{
closest_line_sine(bp = NULL, genome = NULL, cores = 1)
}
\arguments{
\item{bp}{\href{https://bedtools.readthedocs.io/en/latest/content/general-usage.html#bedpe-format}{Bedpe} from \link[InProgress]{svaba_vcf2bedpe} or \link[InProgress]{manta_vcf2bedpe}}

\item{cores}{Number of cores to run on, default is 1}
}
\value{
\href{https://bedtools.readthedocs.io/en/latest/content/general-usage.html#bedpe-format}{Bedpe} with a columns added for distance to nearest LINE and SINE element
}
\description{
Annotates the distance to nearest LINE and SINE element using repeat masker elements from \href{http://www.repeatmasker.org/genomicDatasets/RMGenomicDatasets.html}{RepeatMasker}.
Annotation is in hg38.
}
\details{
Wrapper for LINE/SINE annotation
}
