% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_sine_annotation.R
\name{find_closest_match_line}
\alias{find_closest_match_line}
\title{annotates the closest line element}
\usage{
find_closest_match_line(i, bedpe_l, LINE_dt = NULL)
}
\arguments{
\item{i}{iteration value}

\item{bedpe_l}{ordered bedpe from \link[InProgress]{check_reformat}}

\item{LINE_dt}{line elements}
}
\value{
data.table with closest line element distance
}
\description{
Annotates the distance to closest line element
}
\details{
Closest line element
}
\keyword{internal}
