% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_sine_annotation.R
\name{find_closest_match_sine}
\alias{find_closest_match_sine}
\title{annotates the closest sine element}
\usage{
find_closest_match_sine(i, bedpe_s)
}
\arguments{
\item{i}{iteration value}

\item{bedpe_s}{ordered bedpe from \link[InProgress]{check_reformat}}
}
\value{
data.table with closest sine element distance
}
\description{
Annotates the distance to closest sine element
}
\details{
Closest sine element
}
\keyword{internal}
