% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_annotation.R
\name{find_gene_overlap}
\alias{find_gene_overlap}
\title{SV occurance in genes}
\usage{
find_gene_overlap(bed_g = NULL)
}
\arguments{
\item{bed_g}{bedpe row}
}
\value{
\code{GRanges} of genes
}
\description{
Determines if an SV is overlapping any genes
}
\details{
Gene overlap
}
\keyword{internal}
