% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_annotation.R
\name{funct_annot}
\alias{funct_annot}
\title{Annotates SVs}
\usage{
funct_annot(bp = NULL, cores = 1)
}
\arguments{
\item{bp}{bedpe file to annotate}

\item{cores}{number of cores to run on, default is 1}
}
\value{
bedpe with funct_annot column
}
\description{
Annotates for introns, coding, cn, inv, non coding, sv impacts
}
