% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_filter.R
\name{fuzzy_filter_germline}
\alias{fuzzy_filter_germline}
\title{Distance to closest germline annotator}
\usage{
fuzzy_filter_germline(itter = NULL, bed = NULL)
}
\arguments{
\item{itter}{passed from mclapply to iterate}

\item{bed}{Bedpe returned from annotate_sv function}
}
\value{
SV data table with columns added indicating germline or somatic, germline is defined as <=1kbp away from agnostic perfect match in reference
}
\description{
Determines if each SV should be considered germline by hard filtering. Used with wrapper function.
}
\keyword{internal}
