% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_filter.R
\docType{data}
\name{gnomad_germline_hg19all}
\alias{gnomad_germline_hg19all}
\title{gNOMAD dataset to fuzzy filter}
\format{
\code{data.table}
}
\usage{
gnomad_germline_hg19all
}
\description{
gNOMAd v2.1 control sites in hg19
}
\keyword{data}
\keyword{internal}
