% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_annotation.R
\docType{data}
\name{hg19_exonlocations}
\alias{hg19_exonlocations}
\title{Ensembl exon coordinates}
\format{
\code{data.table}
}
\usage{
hg19_exonlocations
}
\description{
Exon coordinates hg19
}
\keyword{data}
\keyword{internal}
