% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_annotation.R
\docType{data}
\name{hg19_gencode_genelocations}
\alias{hg19_gencode_genelocations}
\title{Gencode gene coordinates, subset to protein coding}
\format{
\code{data.table}
}
\usage{
hg19_gencode_genelocations
}
\description{
Gene coordinates hg19
}
\keyword{data}
\keyword{internal}
