% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_annotation.R
\docType{data}
\name{hg38_ensembl_exonlocations}
\alias{hg38_ensembl_exonlocations}
\title{Ensembl exon coordinates}
\format{
\code{data.table}
}
\usage{
hg38_ensembl_exonlocations
}
\description{
Exon coordinates hg38
}
\keyword{data}
\keyword{internal}
