% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_annotation.R
\docType{data}
\name{hg38_gencode_genelocations}
\alias{hg38_gencode_genelocations}
\title{Ensembl gene coordinates}
\format{
\code{data.table}
}
\usage{
hg38_gencode_genelocations
}
\description{
Gene coordinates hg38
}
\keyword{data}
\keyword{internal}
