% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_annotation.R
\name{impact_annotation}
\alias{impact_annotation}
\title{Clqssificatiion decision tree for annotation}
\usage{
impact_annotation(i, bed)
}
\arguments{
\item{i}{itteration value}

\item{bed}{file to be itterated over}
}
\value{
row with annotation column
}
\description{
Determines if an SV is Translocation, Non-protein coding, intron, CN, or coding
}
