% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_sine_annotation.R
\docType{data}
\name{LINE_dt_hg19}
\alias{LINE_dt_hg19}
\title{Line repeatmaskers}
\format{
\code{data.table}
}
\usage{
LINE_dt_hg19
}
\description{
repeat masker line elements in hg19
}
\keyword{data}
\keyword{internal}
