% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manta_vcf2bedpe.R
\name{manta_vcf2bedpe}
\alias{manta_vcf2bedpe}
\title{Manta vcf to bedpe converter}
\usage{
manta_vcf2bedpe(path = NULL)
}
\arguments{
\item{path}{file path to vcf}
}
\value{
SV data table in bedpe format
}
\description{
Creates a filtered SV bedpe file from an input vcf
}
\details{
Function to call for vcf to bedpe conversion
}
