% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_sine_annotation.R
\docType{data}
\name{SINE_dt_hg19}
\alias{SINE_dt_hg19}
\title{Sine repeatmaskers}
\format{
\code{data.table}
}
\usage{
SINE_dt_hg19
}
\description{
repeat masker sine elements in hg19
}
\keyword{data}
\keyword{internal}
