% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svaba_vcf2bedpe.R
\name{svaba_vcf2bedpe}
\alias{svaba_vcf2bedpe}
\title{Converts \href{https://github.com/walaj/svaba}{SVABA} VCFs to \href{https://bedtools.readthedocs.io/en/latest/content/general-usage.html#bedpe-format}{Bedpe} format}
\usage{
svaba_vcf2bedpe(filepath = NULL)
}
\arguments{
\item{filepath}{file path to \href{https://github.com/walaj/svaba}{SVABA} VCF}
}
\value{
SV data table in \href{https://bedtools.readthedocs.io/en/latest/content/general-usage.html#bedpe-format}{Bedpe} format
}
\description{
Converts \href{https://github.com/walaj/svaba}{SVABA} VCF to \href{https://bedtools.readthedocs.io/en/latest/content/general-usage.html#bedpe-format}{Bedpe} file format
}
\details{
SVaBa vcf to bedpe
}
