% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_annotation.R
\name{find_exon_overlap}
\alias{find_exon_overlap}
\title{SV occurance in exons}
\usage{
find_exon_overlap(bed_e = NULL)
}
\arguments{
\item{bed_e}{bedpe row}
}
\value{
\code{GRanges} of genes
}
\description{
Determines if an SV is overlapping any genes
}
\details{
Exon overlap
}
\keyword{internal}
