% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svaba_vcf2bedpe.R
\name{svaba_vcf2bedpe}
\alias{svaba_vcf2bedpe}
\title{Converts SVaBa VCFs to bedpe format}
\usage{
svaba_vcf2bedpe(filepath = NULL)
}
\arguments{
\item{file_path:}{file path to SVaBa VCF}
}
\value{
bedpe
}
\description{
Converts SVaBa VCF to bedpe file format
}
\details{
SVaBa vcf to bedpe
}
